Audit for NSInteger and CGFloat

+++ Frameworks/UXKit.framework/Versions/203.38.0/Headers/MissingTypes.h
@@ -21,38 +21,38 @@
 
 @protocol UXTableViewDataSource <NSObject>
 @optional
--(long long)numberOfSectionsInTableView:(id)arg1;
--(id)tableView:(id)arg1 titleForHeaderInSection:(long long)arg2;
--(id)tableView:(id)arg1 titleForFooterInSection:(long long)arg2;
+-(NSInteger)numberOfSectionsInTableView:(id)arg1;
+-(id)tableView:(id)arg1 titleForHeaderInSection:(NSInteger)arg2;
+-(id)tableView:(id)arg1 titleForFooterInSection:(NSInteger)arg2;
 -(BOOL)tableView:(id)arg1 canEditRowAtIndexPath:(id)arg2;
 -(BOOL)tableView:(id)arg1 canMoveRowAtIndexPath:(id)arg2;
 -(id)sectionIndexTitlesForTableView:(id)arg1;
--(long long)tableView:(id)arg1 sectionForSectionIndexTitle:(id)arg2 atIndex:(long long)arg3;
--(void)tableView:(id)arg1 commitEditingStyle:(long long)arg2 forRowAtIndexPath:(id)arg3;
+-(NSInteger)tableView:(id)arg1 sectionForSectionIndexTitle:(id)arg2 atIndex:(NSInteger)arg3;
+-(void)tableView:(id)arg1 commitEditingStyle:(NSInteger)arg2 forRowAtIndexPath:(id)arg3;
 -(void)tableView:(id)arg1 moveRowAtIndexPath:(id)arg2 toIndexPath:(id)arg3;
 
 @required
--(long long)tableView:(id)arg1 numberOfRowsInSection:(long long)arg2;
+-(NSInteger)tableView:(id)arg1 numberOfRowsInSection:(NSInteger)arg2;
 -(id)tableView:(id)arg1 cellForRowAtIndexPath:(id)arg2;
 @end
 
 @protocol UXTableViewDelegate <NSObject>
 @optional
 -(void)tableView:(id)arg1 willDisplayCell:(id)arg2 forRowAtIndexPath:(id)arg3;
--(void)tableView:(id)arg1 willDisplayHeaderView:(id)arg2 forSection:(long long)arg3;
--(void)tableView:(id)arg1 willDisplayFooterView:(id)arg2 forSection:(long long)arg3;
+-(void)tableView:(id)arg1 willDisplayHeaderView:(id)arg2 forSection:(NSInteger)arg3;
+-(void)tableView:(id)arg1 willDisplayFooterView:(id)arg2 forSection:(NSInteger)arg3;
 -(void)tableView:(id)arg1 didEndDisplayingCell:(id)arg2 forRowAtIndexPath:(id)arg3;
--(void)tableView:(id)arg1 didEndDisplayingHeaderView:(id)arg2 forSection:(long long)arg3;
--(void)tableView:(id)arg1 didEndDisplayingFooterView:(id)arg2 forSection:(long long)arg3;
+-(void)tableView:(id)arg1 didEndDisplayingHeaderView:(id)arg2 forSection:(NSInteger)arg3;
+-(void)tableView:(id)arg1 didEndDisplayingFooterView:(id)arg2 forSection:(NSInteger)arg3;
 -(double)tableView:(id)arg1 heightForRowAtIndexPath:(id)arg2;
--(double)tableView:(id)arg1 heightForHeaderInSection:(long long)arg2;
--(double)tableView:(id)arg1 heightForFooterInSection:(long long)arg2;
+-(double)tableView:(id)arg1 heightForHeaderInSection:(NSInteger)arg2;
+-(double)tableView:(id)arg1 heightForFooterInSection:(NSInteger)arg2;
 -(double)tableView:(id)arg1 estimatedHeightForRowAtIndexPath:(id)arg2;
--(double)tableView:(id)arg1 estimatedHeightForHeaderInSection:(long long)arg2;
--(double)tableView:(id)arg1 estimatedHeightForFooterInSection:(long long)arg2;
--(id)tableView:(id)arg1 viewForHeaderInSection:(long long)arg2;
--(id)tableView:(id)arg1 viewForFooterInSection:(long long)arg2;
--(long long)tableView:(id)arg1 accessoryTypeForRowWithIndexPath:(id)arg2;
+-(double)tableView:(id)arg1 estimatedHeightForHeaderInSection:(NSInteger)arg2;
+-(double)tableView:(id)arg1 estimatedHeightForFooterInSection:(NSInteger)arg2;
+-(id)tableView:(id)arg1 viewForHeaderInSection:(NSInteger)arg2;
+-(id)tableView:(id)arg1 viewForFooterInSection:(NSInteger)arg2;
+-(NSInteger)tableView:(id)arg1 accessoryTypeForRowWithIndexPath:(id)arg2;
 -(void)tableView:(id)arg1 accessoryButtonTappedForRowWithIndexPath:(id)arg2;
 -(BOOL)tableView:(id)arg1 shouldHighlightRowAtIndexPath:(id)arg2;
 -(void)tableView:(id)arg1 didHighlightRowAtIndexPath:(id)arg2;
@@ -61,14 +61,14 @@
 -(id)tableView:(id)arg1 willDeselectRowAtIndexPath:(id)arg2;
 -(void)tableView:(id)arg1 didSelectRowAtIndexPath:(id)arg2;
 -(void)tableView:(id)arg1 didDeselectRowAtIndexPath:(id)arg2;
--(long long)tableView:(id)arg1 editingStyleForRowAtIndexPath:(id)arg2;
+-(NSInteger)tableView:(id)arg1 editingStyleForRowAtIndexPath:(id)arg2;
 -(id)tableView:(id)arg1 titleForDeleteConfirmationButtonForRowAtIndexPath:(id)arg2;
 -(id)tableView:(id)arg1 editActionsForRowAtIndexPath:(id)arg2;
 -(BOOL)tableView:(id)arg1 shouldIndentWhileEditingRowAtIndexPath:(id)arg2;
 -(void)tableView:(id)arg1 willBeginEditingRowAtIndexPath:(id)arg2;
 -(void)tableView:(id)arg1 didEndEditingRowAtIndexPath:(id)arg2;
 -(id)tableView:(id)arg1 targetIndexPathForMoveFromRowAtIndexPath:(id)arg2 toProposedIndexPath:(id)arg3;
--(long long)tableView:(id)arg1 indentationLevelForRowAtIndexPath:(id)arg2;
+-(NSInteger)tableView:(id)arg1 indentationLevelForRowAtIndexPath:(id)arg2;
 -(BOOL)tableView:(id)arg1 shouldShowMenuForRowAtIndexPath:(id)arg2;
 -(BOOL)tableView:(id)arg1 canPerformAction:(SEL)arg2 forRowAtIndexPath:(id)arg3 withSender:(id)arg4;
 -(void)tableView:(id)arg1 performAction:(SEL)arg2 forRowAtIndexPath:(id)arg3 withSender:(id)arg4;
