+++ Frameworks/UXKit.framework/Versions/203.38.0/Headers/NSWindow-UXKitAppearance.h
@@ -9,7 +9,7 @@
 
 @interface NSWindow (UXKitAppearance) <UXKitAppearance>
 - (void)tintColorDidChange;
-@property (nonatomic) long long tintAdjustmentMode;
+@property (nonatomic) NSInteger tintAdjustmentMode; // TODO: enum
 @property (nonatomic, strong) NSColor *tintColor;
 
 @end
