+++ Frameworks/UXKit.framework/Versions/203.38.0/Headers/UXBarPositioning-Protocol.h
@@ -7,6 +7,6 @@
 
 
 @protocol UXBarPositioning <NSObject>
-@property (nonatomic, readonly) long long barPosition;
+@property (nonatomic, readonly) NSInteger barPosition; // TODO: enum
 @end
