+++ Frameworks/UXKit.framework/Versions/203.38.0/Headers/UXBarPositioningDelegate-Protocol.h
@@ -9,6 +9,6 @@
 @protocol UXBarPositioningDelegate <NSObject>
 
 @optional
-- (long long)positionForBar:(id <UXBarPositioning>)arg1;
+- (NSInteger)positionForBar:(id <UXBarPositioning>)arg1; // TODO: enum
 @end
