+++ Frameworks/UXKit.framework/Versions/203.38.0/Headers/UXClickEventTracker.h
@@ -8,7 +8,7 @@
 
 @interface UXClickEventTracker : UXEventTracker
 
-@property (nonatomic) unsigned long long numberOfClicksRequired;
+@property (nonatomic) NSUInteger numberOfClicksRequired;
 - (void)mouseUp:(id)arg1;
 - (void)mouseDown:(id)arg1;
