+++ Frameworks/UXKit.framework/Versions/203.38.0/Headers/UXCollectionViewMutableIndexPathsSet.h
@@ -9,19 +9,19 @@
 @interface UXCollectionViewMutableIndexPathsSet : UXCollectionViewIndexPathsSet
 
 - (void)adjustForDeletionOfIndexPath:(id)arg1;
-- (void)adjustForDeletionOfItems:(id)arg1 inSection:(unsigned long long)arg2;
+- (void)adjustForDeletionOfItems:(id)arg1 inSection:(NSUInteger)arg2;
 - (void)adjustForInsertionOfIndexPath:(id)arg1;
-- (void)adjustForInsertionOfItems:(id)arg1 inSection:(unsigned long long)arg2;
+- (void)adjustForInsertionOfItems:(id)arg1 inSection:(NSUInteger)arg2;
 - (void)adjustForDeletionOfSections:(id)arg1;
-- (void)adjustForDeletionOfSection:(unsigned long long)arg1;
+- (void)adjustForDeletionOfSection:(NSUInteger)arg1;
 - (void)adjustForInsertionOfSections:(id)arg1;
-- (void)adjustForInsertionOfSection:(unsigned long long)arg1;
+- (void)adjustForInsertionOfSection:(NSUInteger)arg1;
 - (void)intersectIndexPathsSet:(id)arg1;
 - (void)removeAllIndexPaths;
 - (void)removeIndexPathsSet:(id)arg1;
 - (void)removeIndexPaths:(id)arg1;
 - (void)removeIndexPath:(id)arg1;
-- (void)addSection:(long long)arg1 itemsInRange:(NSRange)arg2;
+- (void)addSection:(NSInteger)arg1 itemsInRange:(NSRange)arg2;
 - (void)addIndexPathsSet:(id)arg1;
 - (void)addIndexPaths:(id)arg1;
 - (void)addIndexPath:(id)arg1;
